Revised checkout endpoint (#65)
* made changes to the checkout service implementation and coverted Object types oF Double to Big integer

* resolved merge conflict

* changes made

* changes made

* comments removed

+++ src/main/java/com/example/safariwebstore008/dto/ProcessPaymentDTO.java
@@ -3,11 +3,13 @@ package com.example.safariwebstore008.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 public class ProcessPaymentDTO {
     private String firstName;
     private String lastName;
     private String email;
-    private Double amount;
+    private BigInteger amount;
 }
